creating packages/core-types to avoid duplication across app silos
