+++ apps/backend/package.json
@@ -30,6 +30,7 @@
     "@hono/zod-openapi": "^1.1.1",
     "@prisma/client": "^6.13.0",
     "@semiont/api-types": "*",
+    "@semiont/core-types": "*",
     "dotenv": "^16.3.1",
     "google-auth-library": "^9.10.0",
     "gremlin": "^3.7.4",
